requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf
google-genai
